Extract Prompt Studio advanced field UI
